fix(core): carry the bifur-client connection fix as a pnpm patch until it is released
