+++ patches/@sanity__bifur-client@1.0.0.patch
@@ -1,0 +1,141 @@
+diff --git a/dist/index.d.ts b/dist/index.d.ts
+index 793e01fd9d183c1761081dc29bb24e92295fb008..b9160620cb8c68f738ba205b2824762e8775957c 100644
+--- a/dist/index.d.ts
++++ b/dist/index.d.ts
+@@ -112,4 +112,23 @@ export declare interface SanityClientLike {
+  */
+ export declare type SubscribeMethods = 'presence' | 'listen' | 'authorization'
+ 
++/**
++ * The error emitted when the WebSocket connection fails or is closed by the
++ * other end. `code` and `reason` are only set for `CONNECTION_CLOSED` errors,
++ * mirroring the `CloseEvent` that produced them.
++ *
++ * @public
++ */
++export declare class WebSocketError extends Error {
++  type: 'CONNECTION_ERROR' | 'CONNECTION_CLOSED'
++  code: number | undefined
++  reason: string | undefined
++  constructor(
++    message: string,
++    type: 'CONNECTION_ERROR' | 'CONNECTION_CLOSED',
++    code?: number,
++    reason?: string,
++  )
++}
++
+ export {}
+diff --git a/dist/index.js b/dist/index.js
+index 41be66be926d88cdaad96223a955f0925e8cce35..895ad1f9df025499e15a3c34d8976f0b27337e18 100644
+--- a/dist/index.js
++++ b/dist/index.js
+@@ -1,5 +1,12 @@
++// PATCHED via pnpm patch: applies the connection fix from
++// https://github.com/sanity-io/bifur-client/pull/30 ahead of its release —
++// remove the patch (and this file's changes) when bumping to the version that
++// includes it. One deliberate divergence from that PR as currently written:
++// the disconnect delay is a wall-clock grace period (timer(5e3)) instead of
++// two chained setTimeout(0) hops, which measurably still churn connections
++// during studio boot (see the share() comment below).
+ import { switchMap, mergeMap, share, distinctUntilChanged, take, map, shareReplay, filter, finalize, takeUntil } from "rxjs/operators";
+-import { partition, fromEvent, EMPTY, of, combineLatest, merge, defer, Observable, race, timer, throwError, NEVER } from "rxjs";
++import { partition, fromEvent, EMPTY, of, combineLatest, merge, defer, Observable, race, timer, throwError, NEVER, ReplaySubject } from "rxjs";
+ import { customAlphabet } from "nanoid";
+ const defaultGetNextRequestId = customAlphabet(
+   "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz-",
+@@ -103,18 +110,29 @@ const createClient = (connection$, options = {}) => {
+     request: (method, params) => requestMethod(method, params)
+   };
+ };
++const CLOSE_CODE_NORMAL = 1e3, CLOSE_REASON = "WebSockets connection closed by client";
+ class WebSocketError extends Error {
+   type;
+   code;
+   reason;
+   constructor(message, type, code, reason) {
+-    super(message), this.type = type, this.code = code, this.reason = reason;
++    super(message), this.name = "WebSocketError", this.type = type, this.code = code, this.reason = reason;
+   }
+ }
++// Closes the socket without triggering the "WebSocket is closed before the
++// connection is established" browser warning: calling `close()` on a socket
++// that is still `CONNECTING` aborts the handshake and logs that warning, so
++// instead the close is deferred until the handshake settles. If the handshake
++// fails, the socket never opens and there is nothing to close.
++function closeSocket(ws) {
++  ws.readyState === ws.CONNECTING ? ws.onopen = () => ws.close(CLOSE_CODE_NORMAL, CLOSE_REASON) : ws.readyState === ws.OPEN && ws.close(CLOSE_CODE_NORMAL, CLOSE_REASON);
++}
++function detachHandlersAndClose(ws) {
++  ws.onopen = null, ws.onerror = null, ws.onclose = null, closeSocket(ws);
++}
+ function createConnect(getWebsocketInstance) {
+   return (url) => new Observable((subscriber) => {
+     const ws = getWebsocketInstance(url);
+-    let didUnsubscribe = !1;
+     const onOpen = () => {
+       subscriber.next(ws);
+     }, onError = () => {
+@@ -122,7 +140,7 @@ function createConnect(getWebsocketInstance) {
+         new WebSocketError("WebSocket connection error", "CONNECTION_ERROR")
+       );
+     }, onClose = (ev) => {
+-      didUnsubscribe ? subscriber.complete() : subscriber.error(
++      subscriber.error(
+         new WebSocketError(
+           "WebSocket connection error",
+           "CONNECTION_CLOSED",
+@@ -132,7 +150,7 @@ function createConnect(getWebsocketInstance) {
+       );
+     };
+     return ws.onopen = onOpen, ws.onclose = onClose, ws.onerror = onError, () => {
+-      didUnsubscribe = !0, ws.close(1e3, "WebSockets connection closed by client");
++      detachHandlersAndClose(ws);
+     };
+   });
+ }
+@@ -162,11 +180,31 @@ function fromUrl(url, options = {}) {
+           )
+         )
+       ) : id,
+-      shareReplay({ refCount: !0 }),
++      // Ensure graceful disconnect in browsers. This sits before the `share`
++      // so that unloading the page completes through it and closes the socket
++      // immediately, rather than waiting on the disconnect timer.
+       takeUntil(
+-        // ensure graceful disconnect in browsers
+         isEventTargetLike(ourGlobal) ? fromEvent(ourGlobal, "beforeunload") : NEVER
+-      )
++      ),
++      // Share a single connection between all subscribers, disconnecting only
++      // after a wall-clock grace period without any. This makes the connection
++      // compatible with the teardown scheduling of consumers like react-rx's
++      // `useObservable`: it warms up an uncached observable with a render-phase
++      // subscribe/unsubscribe whose shared store resets one microtask later,
++      // while the real subscription only arrives with React's passive effects.
++      // Under a loaded main thread (studio boot) that effect flush lands
++      // 100-200ms later — task-count based delays (one or two setTimeout hops,
++      // as bifur-client#30 currently uses) measurably still tear the connection
++      // down in between, reconnecting two to three times per page load. Only a
++      // wall-clock grace absorbs the gap; cold mounts then reuse the in-flight
++      // connection instead of aborting the handshake ("WebSocket is closed
++      // before the connection is established") and reconnecting.
++      share({
++        connector: () => new ReplaySubject(1),
++        resetOnError: !0,
++        resetOnComplete: !0,
++        resetOnRefCountZero: () => timer(5e3)
++      })
+     ),
+     { token$ }
+   );
+@@ -183,8 +221,8 @@ function fromSanityClient(client) {
+ }
+ export {
+   ERROR_CODES,
++  WebSocketError,
+   createClient,
+   fromSanityClient,
+   fromUrl
+ };
+-//# sourceMappingURL=index.js.map
